+++ src/test/compile-fail/trait-impl-different-num-params.rs
@@ -1,4 +1,4 @@
-// Copyright 2012 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2012-2014 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -9,11 +9,11 @@
 // except according to those terms.
 
 trait foo {
-    fn bar(&self, x: uint) -> Self;
+    fn bar(&self, x: usize) -> Self;
 }
-impl foo for int {
-    fn bar(&self) -> int {
-        //~^ ERROR method `bar` has 0 parameters but the trait has 1
+impl foo for isize {
+    fn bar(&self) -> isize {
+        //~^ ERROR method `bar` has 1 parameter but the declaration in trait `foo::bar` has 2
         *self
     }
 }
